Refactor dao interfaces to repository
