--- src/main/java/App/dao/CourseDao.java
@@ -1,8 +1,0 @@
-package App.dao;
-
-import App.dao.entity.Course;
-
-public interface CourseDao extends AbstractDao<Long, Course> {
-
-    Course getByName(String name);
-}
