--- src/main/java/App/dao/OrderDao.java
@@ -1,7 +1,0 @@
-package App.dao;
-
-
-import App.dao.entity.Order;
-
-public interface OrderDao extends AbstractDao<Long, Order> {
-}
