--- src/main/java/App/dao/OrderInfoDao.java
@@ -1,10 +1,0 @@
-package App.dao;
-
-import App.dao.entity.OrderInfo;
-
-import java.util.List;
-
-public interface OrderInfoDao extends AbstractDao<Long, OrderInfo> {
-
-    List<OrderInfo> getByOrderId(Long id);
-}
