+++ src/main/java/App/repository/CourseRepository.java
@@ -1,0 +1,8 @@
+package App.repository;
+
+import App.repository.entity.Course;
+
+public interface CourseRepository extends AbstractRepository<Long, Course> {
+
+    Course getByName(String name);
+}
