+++ src/main/java/App/repository/OrderRepository.java
@@ -1,0 +1,7 @@
+package App.repository;
+
+
+import App.repository.entity.Order;
+
+public interface OrderRepository extends AbstractRepository<Long, Order> {
+}
